Refactor: Make Dir (Activity, Fragment) activity와 fragment 디렉토리 나눔
