+++ .idea/misc.xml
@@ -3,7 +3,32 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/activity_my_farm.xml" value="0.22" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/activity_my_page.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/activity_new_my_farm.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/activity_other.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/activity_search_address.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/activity_search_address2.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/activity_show_image.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/activity_splash.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/add_community_image_view.xml" value="0.1875" />
         <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/cardview_farm.xml" value="0.17806484295845998" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/cardview_setting.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/community_view.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_alarm.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_check_type_for_google.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_community.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_login2.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_main.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_main2.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_map.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_my_page.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_products.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_setting.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/fragment_signup.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/image_view.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/imgview_for_community.xml" value="0.1875" />
+        <entry key="..\:/AndroidStudioProject/NURU/app/src/main/res/layout/viewholder_search_result_item.xml" value="0.1875" />
         <entry key="..\:/Users/Choi/AndroidStudioProjects/NURU/app/src/main/res/drawable/radius_gray.xml" value="0.26564102564102565" />
         <entry key="..\:/Users/Choi/AndroidStudioProjects/NURU/app/src/main/res/drawable/radius_green.xml" value="0.26564102564102565" />
         <entry key="..\:/Users/Choi/AndroidStudioProjects/NURU/app/src/main/res/drawable/radius_white.xml" value="0.26564102564102565" />
@@ -103,7 +128,7 @@
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="11" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="11" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
