+++ app/build.gradle
@@ -6,6 +6,7 @@ plugins {
     id("androidx.navigation.safeargs")
 }
 apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-kapt'
 
 android {
     compileSdk 32
@@ -131,6 +132,9 @@ dependencies {
 
     //pull and refresh
     implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.1.0"
+
+    //databinding
+    kapt 'com.android.databinding:compiler:3.5.3'
 
 }
 repositories {
