+++ app/src/main/java/com/example/nuru/View/Adapter/ImgInCommunityAdapter.kt
@@ -10,7 +10,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.example.nuru.View.Activity.ShowImageActivity
+import com.example.nuru.View.Activity.Community.ShowImageActivity
 import com.example.nuru.databinding.ImgviewForCommunityBinding
 
 class ImgInCommunityAdapter(private val context: Context) :
