+++ app/src/main/res/layout/activity_add_admin.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Activity.AddAdminActivity">
+    tools:context=".View.Activity.MyPage.AddAdminActivity">
 
     <EditText
         android:id="@+id/edt_addminId"
