+++ app/src/main/res/layout/activity_add_community.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Activity.AddCommunityActivity">
+    tools:context=".View.Activity.Community.AddCommunityActivity">
 
     <ScrollView
         android:layout_width="match_parent"
