+++ app/src/main/res/layout/activity_add_farm.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Activity.AddFarmActivity">
+    tools:context=".View.Activity.MyPage.AddFarmActivity">
 
     <TextView
         android:id="@+id/txt_farmAddress"
