+++ app/src/main/res/layout/activity_add_image.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Activity.AddImageActivity">
+    tools:context=".View.Activity.Community.AddImageActivity">
 
     <ImageView
         android:id="@+id/img_UploadImage"
