+++ app/src/main/res/layout/activity_edit_community.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Activity.EditCommunityActivity">
+    tools:context=".View.Activity.Community.EditCommunityActivity">
 
     <ScrollView
         android:layout_width="match_parent"
