+++ app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Activity.LoginActivity">
+    tools:context=".View.Activity.Login.LoginActivity">
 
     <fragment
         android:id="@+id/login_navigation"
