+++ app/src/main/res/layout/activity_maps.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layout_constraintTop_toTopOf="parent"
-        tools:context=".View.Activity.MapsActivity"
+        tools:context=".View.Activity.Map.MapsActivity"
         tools:layout_editor_absoluteX="0dp" />
 
     <ImageButton
