+++ app/src/main/res/layout/activity_maps2.xml
@@ -10,7 +10,7 @@
         android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".View.Activity.MapsActivity"
+        tools:context=".View.Activity.Map.MapsActivity"
         tools:layout_editor_absoluteX="0dp"
         tools:layout_editor_absoluteY="0dp" />
