+++ app/src/main/res/layout/activity_my_farm.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Activity.MyFarmActivity">
+    tools:context=".View.Activity.MyPage.MyFarmActivity">
 
     <ProgressBar
         android:id="@+id/widget_progressBarInMyFarm"
