+++ app/src/main/res/layout/activity_my_page.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Activity.MyPageActivity">
+    tools:context=".View.Activity.MyPage.MyPageActivity">
 <!--
     <fragment
         android:id="@+id/mypage_navigation"
