+++ app/src/main/res/layout/activity_search_address.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Activity.SearchAddressActivity">
+    tools:context=".View.Activity.Map.SearchAddressActivity">
 
     <EditText
         android:id="@+id/et_SearchBarInputView"
