+++ app/src/main/res/layout/activity_show_image.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".View.Activity.ShowImageActivity">
+    tools:context=".View.Activity.Community.ShowImageActivity">
 
 
     <androidx.viewpager2.widget.ViewPager2
