+++ app/src/main/res/layout/fragment_setting.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".View.Fragment.SettingFragment">
+    tools:context=".View.Fragment.MyPage.SettingFragment">
 
     <TextView
         android:id="@+id/txt_Setting"
