+++ app/src/main/res/layout/fragment_signup.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Fragment.SignupFragment">
+    tools:context=".View.Fragment.Login.SignupFragment">
     <ScrollView
         android:layout_width="fill_parent"
         android:layout_height="fill_parent"
