+++ app/src/main/res/navigation/login_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment2"
-        android:name="com.example.nuru.View.Fragment.LoginFragment"
+        android:name="com.example.nuru.View.Fragment.Login.LoginFragment"
         android:label="fragment_login2"
         tools:layout="@layout/fragment_login2" >
         <action
@@ -25,7 +25,7 @@
     </fragment>
     <fragment
         android:id="@+id/signupFragment"
-        android:name="com.example.nuru.View.Fragment.SignupFragment"
+        android:name="com.example.nuru.View.Fragment.Login.SignupFragment"
         android:label="fragment_signup"
         tools:layout="@layout/fragment_signup" >
         <action
@@ -34,17 +34,17 @@
     </fragment>
     <activity
         android:id="@+id/mapsActivity"
-        android:name="com.example.nuru.View.Activity.MapsActivity"
+        android:name="com.example.nuru.View.Activity.Map.MapsActivity"
         android:label="activity_maps"
         tools:layout="@layout/activity_maps" />
     <activity
         android:id="@+id/myPageActivity"
-        android:name="com.example.nuru.View.Activity.MyPageActivity"
+        android:name="com.example.nuru.View.Activity.MyPage.MyPageActivity"
         android:label="activity_my_page"
         tools:layout="@layout/activity_my_page" />
     <fragment
         android:id="@+id/checkTypeForGoogleFragment"
-        android:name="com.example.nuru.View.Fragment.CheckTypeForGoogleFragment"
+        android:name="com.example.nuru.View.Fragment.Login.CheckTypeForGoogleFragment"
         android:label="fragment_check_type_for_google"
         tools:layout="@layout/fragment_check_type_for_google">
         <argument
